requirements.txt.txt
requests>=2.31.0
openai>=1.3.5
python-dotenv>=1.0.0
